pyproject.toml
[project]
name = "claudesavvy"
version = "1.0.6"
description = "Web-based usage monitoring tool for Claude Code"
authors = [{name = "Allan Napier"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "click>=8.1.0",
    "rich>=13.0.0",
    "python-dateutil>=2.8.0",
    "Flask>=3.0.0",
    "Jinja2>=3.1.0",
]

[project.scripts]
claudesavvy = "claudesavvy.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.package-data]
"claudesavvy.web" = ["templates/**/*.html", "static/**/*"]
